Crate enumflags2_derive

Structs

Flag 🔒

Parameters 🔒

Enums

FlagValue 🔒

Functions

check_flag 🔒

Returns deferred checks

collect_flags 🔒

extract_repr 🔒

Given a list of attributes, find the repr, if any, and return the integer type specified.

fold_expr 🔒

Try to evaluate the expression given.

gen_enumflags 🔒

infer_values 🔒

inferred_value 🔒

type_bits 🔒

Check the repr and return the number of bits available

Attribute Macros

bitflags_internal