Crate enumflags2_derive

Structs

Flag ğŸ”’

Parameters ğŸ”’

Enums

FlagValue ğŸ”’

Functions

check_flag ğŸ”’

Returns deferred checks

collect_flags ğŸ”’

extract_repr ğŸ”’

Given a list of attributes, find the repr, if any, and return the integer type specified.

fold_expr ğŸ”’

Try to evaluate the expression given.

gen_enumflags ğŸ”’

infer_values ğŸ”’

inferred_value ğŸ”’

type_bits ğŸ”’

Check the repr and return the number of bits available

Attribute Macros

bitflags_internal